+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Next
+- Add `get` and `set` default commands, which act as a persistent key-value store.
+  - Keys can begin with a dot to be per-session only.
+  - Keys can begin with a $ to be game-wide, shared among all players.
+  - Keys can begin with $. to be game-wide and per-session.
+
 # v1.7.1
 - Handle non-string error values correctly